test(hooks): move alias, scope, and isolation guard tests out of the monolith
Second half of splitting the 2,077-line
tests/hooks/test_push_pr_script_identity_guard.py. Removes the original
and lands the three remaining modules:

  test_push_pr_guard_aliases.py    301  symlinks, aliases, renamed copies
  test_push_pr_guard_scope.py      467  relevance gate: what the guard polices
  test_push_pr_guard_isolation.py  298  plugin root, -I mode, shipped shim

Every guard module is now under 500 lines and no test was renamed,
weakened, or dropped: 951 collected before the split, 951 after, all
passing. The count is the same measurement on both sides, so a silently
dropped test would show up as a smaller number rather than a green run.

Refs #4764

Copilot-Session: 2005b9d4-c29c-460c-8b98-878535ec8604
